# [Dart] Loops optimization