Optimized loops.