@jinyus 2023-10-18
Dart VM (5k is slower but faster overall):

Main branch:
```
137.92 ms
2.42 s
21.09 s
```

This PR:
```
Dart VM:

        Benchmark 1: dart related.dart
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 172ms
        Processing time (w/o IO): 168ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 168ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 131ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 169ms
        Processing time (w/o IO): 169ms
          Time (mean ± σ):     538.2 ms ±  17.1 ms    [User: 646.0 ms, System: 105.1 ms]
          Range (min … max):   498.3 ms … 560.6 ms    10 runs
         
Dart VM:

        Benchmark 1: dart related.dart
        Processing time (w/o IO): 2043ms
        Processing time (w/o IO): 2232ms
        Processing time (w/o IO): 2242ms
          Time (mean ± σ):      2.905 s ±  0.002 s    [User: 3.118 s, System: 0.130 s]
          Range (min … max):    2.904 s …  2.907 s    2 runs
         
Dart VM:

        Benchmark 1: dart related.dart
        Processing time (w/o IO): 19673ms
        Processing time (w/o IO): 17852ms
        Processing time (w/o IO): 19684ms
          Time (mean ± σ):     20.277 s ±  1.280 s    [User: 20.461 s, System: 0.270 s]
          Range (min … max):   19.372 s … 21.182 s    2 runs
```