@jinyus 2023-10-18
Dart AOT (Faster overall)

Main branch:
```
185.23 ms 
2.89 s 
25.80 s
```
```
Dart AOT:

        Benchmark 1: ./related
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 131ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
        Processing time (w/o IO): 130ms
          Time (mean ± σ):     257.5 ms ±   1.6 ms    [User: 240.7 ms, System: 20.8 ms]
          Range (min … max):   255.2 ms … 259.3 ms    10 runs
         
Dart AOT:

        Benchmark 1: ./related
        Processing time (w/o IO): 2006ms
        Processing time (w/o IO): 2006ms
        Processing time (w/o IO): 2001ms
          Time (mean ± σ):      2.490 s ±  0.001 s    [User: 2.443 s, System: 0.066 s]
          Range (min … max):    2.489 s …  2.490 s    2 runs
         
Dart AOT:

        Benchmark 1: ./related
        Processing time (w/o IO): 17893ms
        Processing time (w/o IO): 17833ms
        Processing time (w/o IO): 17827ms
          Time (mean ± σ):     19.334 s ±  0.023 s    [User: 19.316 s, System: 0.163 s]
          Range (min … max):   19.317 s … 19.351 s    2 runs
```